Add CMake support for build and refactor the directory
It's hard for building using `make`, although it may be simple but
cannot be maintained simply. So this commits use `CMake` for supporting
the whole project building.

It's difficult to use `NASM` for `CMake`. The most strange thing is that
`set(CMAKE_ASM_NASM_SOURCE_FILE_EXTENSIONS s S)` doesn't work, so change
all the assembly file subffix to `.nasm`.
See https://stackoverflow.com/questions/56420035/how-to-use-cmake-to-build-binaries-with-nasm/

Last, use `add_custom_command` for supporting automatically writting to
the hard disk `POST_BUILD`.